ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.7.1
torchvision==0.22.1
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
pydantic==2.5.0
pillow==10.1.0
redis==5.0.1
python-multipart==0.0.6
httpx==0.25.2
numpy==1.26.2
omegaconf==2.3.0
safetensors==0.4.1
huggingface-hub==0.19.4
protobuf>=3.20.0
python-dotenv==1.0.0